// packages/common/autocode/src/lib/enums/ac-enum-condition-operator.enum.ts
export enum AcEnumConditionOperator {
  Between = "between",
  Contains = "contains",
  EndsWith = "ends_with",
  EqualTo = "equal_to",
  GreaterThan = "greater_than",
  GreaterThanEqualTo = "greater_than_equal_to",
  In = "in",
  IsEmpty = "is_empty",
  IsNotNull = "is_not_null",
  IsNull = "is_null",
  LessThan = "less_than",
  LessThanEqualTo = "less_than_equal_to",
  NotIn = "not_in",
  NotEqualTo = "not_equal_to",
  StartsWith = "starts_with",
  Unknown = "unknown",
}